Copy-paste the content of connection.ts and remove the file
